Fresnel Diffraction Rectangular Aperture
Fresnel diffraction plot of the intensity of the transmitted wave when an incident wave diffracts through a rectangular aperture. The plot is of the intensity shown on the screen in the x-y plane. The equation used to find the transmitted field and intensity was taken out of the Goodman's book on Optics, which was the solution to the Fresnel Diffraction Integral.